+++ src-tauri/src/inject/style.js
@@ -68,16 +68,6 @@ window.addEventListener("DOMContentLoaded", (_event) => {
       padding-top: 48px;
     }
 
-    .chakra-ui-light #app .chakra-heading,
-    .chakra-ui-dark #app .chakra-heading,
-    .chakra-ui-light #app .chakra-stack,
-    .chakra-ui-dark #app .chakra-stack,
-    .app-main .sidebar-mouse-in-out,
-    .chakra-modal__content-container .chakra-modal__header > div > div,
-    #__next > div.PageWithSidebarLayout_centeringDiv___L9br > section > header {
-      padding-top: 10px;
-    }
-
     #__next .overflow-hidden>.hidden.bg-gray-900 span.rounded-md.bg-yellow-200 {
       display: none;
     }
@@ -117,187 +107,6 @@ window.addEventListener("DOMContentLoaded", (_event) => {
       display: none;
     }
 
-    #react-root [data-testid="placementTracking"] article,
-    #react-root a[href*="quick_promote_web"],
-    #react-root a[href*="/lists"][role="link"][aria-label],
-    #react-root a[href*="/i/communitynotes"][role="link"][aria-label],
-    #react-root a[role="link"][aria-label="Communities"],
-    #react-root a[role="link"][aria-label="Premium"],
-    #react-root a[role="link"][aria-label="SuperGrok"],
-    #react-root a[href*="/i/verified-orgs-signup"][role="link"][aria-label] {
-      display: none !important;
-    }
-
-    #react-root [data-testid="DMDrawer"],
-    #react-root [data-testid="GrokDrawer"],
-    #react-root [data-testid="chat-drawer-root"],
-    #root > main > footer.justify-center.ease-in {
-      visibility: hidden !important;
-    }
-
-    #__next > div.overflow-hidden.w-full.h-full .absolute.bottom-0.left-0.w-full > div.text-center.text-xs {
-      visibility: hidden !important;
-      height: 0px !important;
-    }
-
-    #react-root [data-testid="primaryColumn"] > div > div {
-      position: relative !important;
-    }
-
-    #react-root [data-testid="sidebarColumn"] {
-      visibility: hidden !important;
-      width: 0 !important;
-      margin: 0 !important;
-      padding: 0 !important;
-      z-index: 1 !important;
-    }
-
-    @media only screen and (min-width: 1000px) {
-      #react-root main[role="main"] {
-        align-items: center !important;
-        overflow-x: clip !important;
-      }
-
-      #react-root [data-testid="primaryColumn"] {
-        width: 700px !important;
-        max-width: 700px !important;
-        margin: 0 auto !important;
-      }
-      #react-root [data-testid="primaryColumn"] > div > div:last-child,
-      #react-root [data-testid="primaryColumn"] > div > div:last-child div {
-        max-width: unset !important;
-      }
-
-      #react-root div[aria-label][role="group"][id^="id__"] {
-        margin-right: 81px !important;
-      }
-
-      #react-root header[role="banner"] {
-        position: fixed !important;
-        left: 0 !important;
-      }
-
-      #react-root header[role="banner"] > div > div > div {
-        justify-content: center !important;
-        padding-top: 0;
-        overflow-x: hidden;
-      }
-
-      #react-root form[role="search"] > div:nth-child(1) > div {
-        background-color: transparent !important;
-      }
-
-      #react-root h1[role="heading"] {
-        padding-top: 4px !important;
-      }
-
-      #react-root header[role="banner"]
-        nav[role="navigation"]
-        *
-        div[dir="auto"]:not([aria-label])
-        > span,
-      #react-root [data-testid="SideNav_AccountSwitcher_Button"] > div:not(:first-child) {
-        display: inline-block !important;
-        opacity: 0 !important;
-        transition: 0.5s cubic-bezier(0.2, 0.8, 0.2, 1);
-      }
-      #react-root header[role="banner"]
-        nav[role="navigation"]:hover
-        *
-        div[dir="auto"]:not([aria-label])
-        > span,
-      #react-root [data-testid="SideNav_AccountSwitcher_Button"]:hover > div:not(:first-child) {
-        opacity: 1 !important;
-      }
-      #react-root header[role="banner"] nav[role="navigation"]:hover > * > div {
-        backdrop-filter: blur(12px) !important;
-      }
-      #react-root header[role="banner"] nav[role="navigation"] > a {
-        position: relative;
-      }
-
-      #react-root header[role="banner"] nav[role="navigation"] > a::before {
-        content: "";
-        position: absolute;
-        top: 0px;
-        right: -40px;
-        bottom: 0px;
-        left: 0px;
-      }
-      #react-root [data-testid="SideNav_AccountSwitcher_Button"] {
-        bottom: 18px !important;
-        left: 1px !important;
-      }
-
-      #react-root [data-testid="SideNav_NewTweet_Button"], #react-root [aria-label="Twitter Blue"]{
-        display: none;
-      }
-    }
-
-    @media only screen and (min-width: 1000px) and (max-width: 1264px) {
-      #react-root [data-testid="sidebarColumn"] form[role="search"] {
-        visibility: visible !important;
-        position: fixed !important;
-        top: 12px !important;
-        right: 16px !important;
-      }
-
-      #react-root [data-testid="sidebarColumn"] input[placeholder="Search"] {
-        width: 150px;
-      }
-
-      #react-root [data-testid="sidebarColumn"] form[role="search"]:focus-within {
-        width: 280px !important;
-        backdrop-filter: blur(12px) !important;
-      }
-
-      #react-root [data-testid="sidebarColumn"] input[placeholder="Search"]:focus {
-        width: 234px !important;
-      }
-    }
-
-    @media only screen and (min-width: 1265px) {
-      #react-root [data-testid="sidebarColumn"] form[role="search"] {
-        visibility: visible !important;
-        position: fixed !important;
-        top: 12px !important;
-        right: 16px !important;
-      }
-
-      #react-root [data-testid="sidebarColumn"] input[placeholder="Search"] {
-        width: 150px;
-      }
-
-      #react-root [data-testid="sidebarColumn"] form[role="search"]:focus-within {
-        width: 374px !important;
-        backdrop-filter: blur(12px) !important;
-      }
-
-      #react-root [data-testid="sidebarColumn"] input[placeholder="Search"]:focus {
-        width: 328px !important;
-      }
-
-      #react-root div[style*="left: -12px"] {
-        left: unset !important;
-      }
-
-      #react-root div[style="left: -8px; width: 306px;"] {
-        left: unset !important;
-        width: 374px !important;
-      }
-
-      #react-root .searchFilters {
-        visibility: visible !important;
-        position: fixed;
-        top: 12px;
-        right: 16px;
-        width: 240px;
-      }
-      #react-root .searchFilters > div > div:first-child {
-        display: none;
-      }
-    }
-
     @media (min-width:1280px){
       #__next .text-base.xl\\:max-w-3xl, #__next form.stretch.xl\\:max-w-3xl {
         max-width: 48rem;
@@ -330,25 +139,13 @@ window.addEventListener("DOMContentLoaded", (_event) => {
 
   // Top spacing adapts to head-hiding scenarios
   const topPaddingCSS = `
-    #layout > ytmusic-nav-bar{
-      padding-top: 20px;
-    }
-
     .columns .column #header,
     .main > div > div.panel.give_me > div.header {
       padding-top: 30px;
     }
 
-    ytd-masthead>#container.style-scope.ytd-masthead {
-      padding-top: 12px;
-    }
-
     #__next header.HeaderBar_header__jn5ju {
       padding-top: 16px;
-    }
-
-    #root > .excalidraw-app> .excalidraw-container .App-menu.App-menu_top{
-      margin-top: 15px;
     }
 
     .geist-page nav.dashboard_nav__PRmJv,
@@ -436,24 +233,8 @@ window.addEventListener("DOMContentLoaded", (_event) => {
       height: 70px;
     }
 
-    .lark > .dashboard-sidebar, .lark > .dashboard-sidebar > .sidebar-user-info , .lark > .dashboard-sidebar .index-module_wrapper_F-Wbq{
-      padding-top:15px;
-    }
-
-    #app-root .mat-mdc-tooltip-trigger.main-menu-button.mdc-icon-button {
-      margin-top: 15px;
-    }
-
-    .lark > .main-wrapper [data-testid="aside"] {
-      top: 15px;
-    }
-
     #global > div.header-container > .mask-paper {
       padding-top: 20px;
-    }
-
-    #background.ytd-masthead {
-      height: 68px;
     }
 
     .wrap.h1body-exist.max-container > div.menu-tocs > div.menu-btn{